Allow .Net Standard apps (aspnetcore) to use applicationinsights.config
